renamed types to typing and some minor styling
