+++ linox/_arithmetic.py
@@ -27,7 +27,7 @@ import plum  # type: ignore  # noqa: PGH003
 
 from linox import utils
 from linox._linear_operator import LinearOperator
-from linox.types import ScalarLike, ShapeLike
+from linox.typing import ScalarLike, ShapeLike
 
 ArithmeticType = LinearOperator | jax.Array
