+++ linox/utils.py
@@ -7,9 +7,7 @@ import jax
 import jax.numpy as jnp
 
 from linox._linear_operator import LinearOperator
-
-# from linox._matrix import Matrix, Scalar
-from linox.types import ArrayLike, DTypeLike, ScalarLike, ShapeLike, ShapeType
+from linox.typing import ArrayLike, DTypeLike, ScalarLike, ShapeLike, ShapeType
 
 __all__ = ["_broadcast_shapes", "as_linop", "as_scalar", "as_shape"]
